Extract methods for dealing with Wikitext
