+++ tests/apis/test_wikimedia.py
@@ -143,29 +143,6 @@ def test_handles_no_matching_languages(wikimedia_api: WikimediaApi) -> None:
     assert wikimedia_api.find_matching_languages(query="ymr") == []
 
 
-def test_get_wikitext(wikimedia_api: WikimediaApi) -> None:
-    actual = wikimedia_api.get_wikitext(
-        filename="File:The main Flickr photo storage server.jpg"
-    )
-
-    with open(
-        "tests/fixtures/wikitext/The_main_Flickr_photo_storage_server.html"
-    ) as infile:
-        expected = infile.read()
-
-    assert actual == expected
-
-
-def test_get_wikitext_for_missing_file(wikimedia_api: WikimediaApi) -> None:
-    with pytest.raises(MissingFileException):
-        wikimedia_api.get_wikitext(filename="File:DefinitelyDoesNotExist.jpg")
-
-
-def test_bad_filename_is_exception(wikimedia_api: WikimediaApi) -> None:
-    with pytest.raises(UnknownWikimediaApiException):
-        wikimedia_api.get_wikitext(filename="File:")
-
-
 def test_get_file_without_structured_data_is_empty(wikimedia_api: WikimediaApi) -> None:
     """
     If a file exists on Commons but it doesn't have any structured data,
